Add experimental support for the Kitty Keyboard Protocol
On startup, GhosttyExtensions.jl emits an escape sequence that asks the terminal to use the Kitty Keyboard Protocol. This allows more flexibility to bind keys. Particularly, the command key (macOS) or hyper key (Linux) can be bound.

Julia's REPL does not understand that protocol and we need to translate some Kitty escape sequences to legacy encodings. This is done by a simple table.